NamespaceAuthenticationManagerTests groovy->java
Issue: gh-4939
